Aligned and styled the logo
Aligned the logo with the help of break tags and flexbox. Added a
gradient effect to the text matching the main background. The css
adds the background color, removes it from around the text (only keeping
the text colored from behind) then makes text transparent and thats how
it works!